refactor: template 생성 함수를 movie(도메인)과 에러로 분리
