+++ index.html
@@ -4,6 +4,7 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <link rel="preload" href="./src/images/workers.png" as="image" />
   <title>영화 리뷰</title>
 </head>
 <body>
